server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apispec==6.6.1
asyncio==3.4.3
attrs==23.2.0
blake3==0.4.1
blinker==1.9.0
botocore==1.23.34
cbor2==5.6.4
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.4
command-not-found==0.3
cryptography==42.0.8
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.17.1
email_validator==2.2.0
falcon==3.1.3
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==22.0.0
hio==0.6.14
hjson==3.1.0
http_sfv==0.9.9
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keri==1.2.0.dev12
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lmdb==1.5.1
lz4==4.4.3
MarkupSafe==2.1.5
marshmallow==3.21.3
mnemonic==0.21
more-itertools==8.10.0
msgpack==1.0.8
multicommand==1.0.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
packaging==24.1
Pillow==9.0.1
pluggy==1.5.0
prettytable==3.10.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pysodium==0.7.17
pytest==8.2.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
PyYAML==6.0.1
qrcode==8.0
referencing==0.35.1
-e git+ssh://git@github.com/GLEIF-IT/reg-pilot-api.git@38edf323bbf1c0ecaa9898f4906bb95de0f54a5f#egg=reg_pilot_api
requests==2.32.3
roman==3.3
rpds-py==0.19.0
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
semver==3.0.2
setuptools==70.3.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlitecloud==0.0.83
starlette==0.46.1
swagger-ui-py==23.9.23
systemd-python==234
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
-e git+ssh://git@github.com/GLEIF-IT/vlei-verifier.git@f6fae4e1c123c7ed3443751ca41b467d64658b4b#egg=verifier
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.43.0
zipp==1.0.0
